upd: adjust the notBefore date of the root by one second.
incrementing its "notBefore" by one second. This accounts for the "infinite chaining" problems in firefox.

Note the nasty hack that is setting not before to a fixed date as openssl has not method to extract the time_t from the ASN1_time back and increment it.
For now this is hardcoded to the notBefore of the current root certificate.